Address more ES inconsistency
It was possible that new charges wouldn't be seen in time to start
adding the charges to accounts. This is addressed by keeping new
charges in memory and passing them directly to the compute charges
function, and then waiting for updates to accounts to become
searchable by using 'refresh = "wait_for"' when bulk uploading
account changes.