Handle case where RequestGpus is None
The GPU credit calculation function was crashing when RequestGpus
was undefined in their job ad because that value turned into None
(rather than being undefined) in the data passed into the function.
Now, instead of only relying on the dict.get() method, there is a
get_clean() function that will return a default value if the output
of ad.get(col, default_value) is None.