HTCONDOR-1677 Create dev environment for CAS
All index names are configurable through the CLI and refer to environment
variables, allowing a pair of activate/deactivate scripts to set up the
env for working with dev indices.

Also added a clone-to-dev script that copies over the account and daily charge
record indices, templates, and aliases to corresponding dev-prepended ES
objects, matching those set by the activate script.