HTCONDOR-1678 CAS accounts should contain both CPU and GPU credits
The current state of the CAS is that CPU accounts are separate from
GPU accounts. This has caused much pain and annoyance for everyone --
developers, administrators, and users -- so this commit aims to
provide both CPU and GPU credit buckets under each account.

This change necessitates some upgrade scripts to bring existing
accounts into compliance and to make sure that new charge records
contain the expected additional information (e.g. jobs are now either
CPU jobs or GPU jobs). Along with this, improvements were made to
setting up dev indices, cloning production indices, and/or backing
up production indices. The setup, upgrade, backup, and cloning scripts
are all under the elasticsearch_templates directory.

The main change to administrators is that they will need to specify
the type of credits being added to an account when creating or upating
an account.

The main change to users will be additional columns in email reports,
separating out CPU and GPU usage. Whether or not to use the CPU or
GPU bucket for a job's charges depends on the existence and value of
the RequestGpus attr (i.e. request_gpus submit file command) in the
job.

An additional change is that project name is now added as an optional
field for accounts. This is not currently used for anything, but could
be useful for storing reference information (e.g. NSF project ID).